tests/requirements.txt
certifi==2025.4.26
charset-normalizer==3.4.2
idna==3.10
iniconfig==2.1.0
packaging==25.0
pluggy==1.6.0
pydicom==2.4.4
pynetdicom==2.1.1
pytest==8.3.5
pytest-dependency==0.6.0
PyYAML==6.0.2
requests==2.32.3
urllib3==2.4.0
xmltodict==0.14.2
